ai-crm-system/requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.9
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
openai-whisper>=20231117
